LayoutTests/imported/w3c/web-platform-tests/web-locks/partitioned-web-locks.tentative.https-expected.txt

< !DOCTYPE html >

PASS WebLocks of an iframe under a 3rd-party site are partitioned
FAIL WebLocks of a nested iframe with a cross-site ancestor are partitioned assert_equals: The 1p iframe failed to acquire the lock expected (undefined) undefined but got (boolean) true
